Move RandomizeWithRegions into pipeline_utils and setup v2
